requirements.txt
grpcio
grpcio-tools
pydantic[email]